don't give an error id of "none" if there is no error report
This was causing db errors trying to do ErrorReport.find("none")

fixes #11816

test plan: configure rails to render the public error pages (turn off
local mode), then add a line to generate a low-level error (such as a
marshal error when reading a cache value, which is how we discovered
this). in the error form the id should be blank, rather than "none".

Change-Id: I83367d1073739bc5588e8cfb59318c502c3eb41a
Reviewed-on: https://gerrit.instructure.com/15367
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>