save new submission version for more actions that can change grades
fixes #11002
fixes #11903

the two cases that can change grades are:
- for gradebook csv uploads
- an assignment value (like points possible) changes

now in both of these cases, a new version of the submission will be saved so
that we have grading history for these changes.  Also, in the case of gradebook
csv upload we are now setting the grader_id to be the user doing the upload.

test plan:
- grade a set of submissions normally
- change the points_possible of the assignment
- look at grading history, it should look right
- now change grades through a csv upload
- look at the grading history, it should look right

Change-Id: I187c8edf74dc9cff83b30d17f1409e92cee359bd
Reviewed-on: https://gerrit.instructure.com/15471
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Stanley Stuart <stanley+gerrit@instructure.com>
Reviewed-by: Amber Taniuchi <amber@instructure.com>