add assignment overrides to ical calendar feeds
This adds assignment override information to the iCal calendar
feed. Student calendar feeds should only contain the latest
override that applies to them. Teachers should see the original
assignment and all the overrides with the override title in
parentheses, similar to how it is displayed in the calendar2
interface.

fixes #11787, #11788, #11789

test plan:
 - create an assignment with an assignment override for a
 course section that has a student
 - subscribe to the student's calendar feed and see the
 assignment event has the section's override date
 - subscribe to the teacher's calendar feed and see both the
 original assignment event and the override events
 - check for regressions on non-overridden assignments

Change-Id: I3cf7eacc40bbc87f693aec463a84670ca9e30063
Reviewed-on: https://gerrit.instructure.com/15641
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Reviewed-by: Marc LeGendre <marc@instructure.com>