Fix Ruby 1.9 Gradebook1 UTF-8 error
In gradebook1 we serialize all columns in User objects, which turns
out to be a bad thing in Ruby 1.9 because it is now UTF-8 aware and
fails for some bad UTF-8 characters in the 'collkey' sortable name
column. We just remove collkey from the serializable columns and
we're good.

Test Plan:
  1. Check that international characters in the sortable name field
     does not break the gradebook1 page.
  2. Specifically, check the original page on beta that discovered
     this bug (and the only one we know for certain reproduces the
     bug):

     https://mtest2.beta-19.instructure.com/courses/8361/gradebook

fixes #CNVS-1972

Change-Id: I3b949faaf4abc4ae0a025ff097ec566c69fa5276
Reviewed-on: https://gerrit.instructure.com/15905
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>