dont force valid email address on mobile login, fixes: #CNVS-2306
test plan:
on an android in the mobile chrome browser go to log into canvas
try to log in with a non-email username (for example, mine is 'ryan')
make sure it lets you log in.

right now it errors because we add "type=email" to
the input to get the nice email-hinted keyboard entry.
but by doing so, it will also use native validatation
to make sure it is a valid email, adding "novalidate"
to the <form> fixes it

Change-Id: Ie3215100d07a932508c366dac147d59502696c8e
Reviewed-on: https://gerrit.instructure.com/16175
QA-Review: Mark Suman <marks@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>