change incoming message processor to prevent mail loops
It looks like we were responding to emails sent by automatic
responders, causing an infinite email loop and flooding
inboxes. Some easy changes to prevent this:

1) If we can't find the original message, then ignore the
incoming message because we don't know what it is and shouldn't
just blindly respond to every incoming message that happens to
match our regex.

2) If the original message wasn't sent because of a
notification, then it's probably a bounce message (or otherwise
unimportant) and we shouldn't respond to the incoming message.

3) Check the incoming message for headers that are commonly
applied to computer generated messages. If we find something to
indicate that the email is an auto-response, don't respond to
it.

fixes #CNVS-2289

test plan:

make sure we don't respond to our own bounce emails:
 - Set a student's calendar notifications to ASAP
 - As a teacher, create a new calendar event for the course
 - View the student's /messages and reply to the calendar
 notification
 - Refresh the page and you should see a bounce message
 - Reply to the bounce message and refresh the page
 - You should not see a new bounce message

make sure legitimate replies are still working:
 - Set a student's discussion topic notifications to ASAP
 - As a teacher, create a new discussion topic in the course
 - Go to the student's /messages and reply to the discussion
 notification
 - The reply should be posted to the discussion

Change-Id: I55d23380f2a20df745e66c06994478ab3694d1bb
Reviewed-on: https://gerrit.instructure.com/16283
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Joel Hough <joel@instructure.com>
QA-Review: Marc LeGendre <marc@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>