Display "Multiple Due Dates" on speedgrader when applicable. Fixes #CNVS-740
On SpeedGrader, when an assignment has multiple due dates, in the header,
display something to let the grader know that.

Testing Notes:
==========
* Using an assignment with no due date, visit SpeedGrader and verify
   that it says, "due: No Due Date"
* Using an assignment with a single (no override) due date, visit
   SpeedGrader and verify that it says "due: <the date>"
* Using an assignment with multiple override due dates, visit
   SpeedGrader and verify that is says "due: Multiple Due Dates"

NOTE: This also changes the previous behavior of not showing anything for a
due date when the assignment didn't have one to now showing
"due: No Due Date" when there is no due date. This is because the assignment
might have no due date but an override might specify one.

Change-Id: I1337015eef623f9579be3af990c8651cacd957af
Reviewed-on: https://gerrit.instructure.com/16391
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
QA-Review: Marc LeGendre <marc@instructure.com>