Use html5 to show videos by default
using a library called mediaElement.js, try to show
all kaltura video and audio using html5 and fall back
on flash if needed.

Test Plan:
 * in all the different browsers (including mobile safari/chrome)
 * try recording webcam videos (as homework submission, as content in a wiki page, etc)
 * view those same pages
 * view recorded submission in speedgrader

use the playlist urls to get kaltura urls

This will get CDN-configured links from Kaltura instead of direct Kaltura
links. It also allows for the case where a MediaObject does not exist. When
one is requested, it will be created in the database and details fetched.

Test Plan for cdn urls:
 * Make sure kaltura is enabled
 * Open your web inspector's network tab
 * Browse to a recorded media file in Canvas, and play it
 * Ensure that the file is streamed from kaltura's configured CDN (s3 in the case of instructure).

Test Plan for nonexistent MediaObjects:
 * Make sure kaltura is enabled
 * In a wiki page, record a media comment
 * Destroy that MediaObject from the database (use Inspect Element to find the media_id which looks like 0_1234abcd)
 * Reload the page, and verify that you can still view the video

fix error when viewing a video file that has an unknown source type

There are some videos that have valid sources that don't
have an extension for some reason. This fixes handling
of those correctly.
There was also a bug that would sort by bitrate as
strings, instead of as integers. The sort function was
refactored out and tests written to ensure that's fixed.

Test Plan
 * find a video in kaltura that has an asset with no extension,
   and make sure you can view that video in a wiki page

refs #CNVS-324

Change-Id: I8ff24a94b8af11fc29b84e45545f8a8b639eeaff
Reviewed-on: https://gerrit.instructure.com/16824
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
QA-Review: Adam Phillipps <adam@instructure.com>