add external tool tag to assignment api
FIXES #CNVS-3625

This attribute was already documented,
but was not actually being included in the
json that the assignment API spits out.

I added the external_tool_tag_attributes
json in cases when the assignment actually
had an external_tool_tag (and had
that as a submission_type).

TEST PLAN:

1) use the API to access an assignment
that has an external_tool_tag

2) you should get "external_tool_tag_attributes"
as an attribute in the assignment json. It should
have the url of the external tool and whether
or not it's a new tab.

Change-Id: Ieebae7df5c169e8c5ae3e31739544014f53c27d3
Reviewed-on: https://gerrit.instructure.com/17509
QA-Review: Clare Hetherington <clare@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>