make conversations specs less prone to failure
Fixes an issue in the conversations selenium specs where the driver was
looking up old element in the page due to the way we were checking for
their existence, and if they did not exist, looking them up.

Also make sure that the elements get refreshed after the message form
submit happens because that appears to modify the dom pretty heavily.

Change-Id: Ib65b9619346acc05d7a69777a6413da286440ee8
Reviewed-on: https://gerrit.instructure.com/17487
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jake Sorce <jake@instructure.com>
QA-Review: Jake Sorce <jake@instructure.com>