accessibility: fix keyboard shortcut dialogs for screen readers
Tested to work in these browsers + screen reader combinations:
  Firefox + JAWS (Windows 7)
  Internet Explorer 9 + JAWS (Windows 7)
  Chrome + VoiceOver (OSX)
  Safari + VoiceOVer (OSX)

Tested and found not to work in these browsers + screen reader
combinations:

  Chrome + JAWS (Windows 7)
  Firefox + VoiceOver (OSX)

Visit the course assignments page and hit the comma key to bring up a
list of keyboard shorcuts. You should be able to navigate the list and
be read the list of keyboard shortcuts in the screen reader. You should
also be able to open and close the dialogue.

When you close the dialogue, you should be taken back to the last
element that had focus.

fixes CNVS-2865

Change-Id: Ieb8b97109a667b9731a35ee1302394eb6d67c21c
Reviewed-on: https://gerrit.instructure.com/17784
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>