fix bug serializing an empty datepicker input
datepicker does not clear the data-date attribute when the input is cleared,
which means that you can't delete a date if you're using this method.

fixes CNVS-4005

test plan:
- with the graded discussion due date field
- set a date, save, edit, clear the date, save
- the date should be cleared

* note: you could also test this with assignment vdd, but there is another bug
that when a override date is cleared it inherits the default date from the
assignment.

Change-Id: Iad16b252d8be9a3c7120db88ef8c6783cd72c6c8
Reviewed-on: https://gerrit.instructure.com/18103
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cameron Matheson <cameron@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>