paginate TokenSelector lists
as part of finishing cross-shard conversations, the rank_results option
to search_messageable_users will no longer be supported (because we
can't calculate ranks in the DB cross-shard, and we're not going to load
the entire collection to sort by rank). this means that we can't rely on
the most useful results being in the first five returned results, so we
actually have to fetch the second page, third page, etc.

to enable this, this commit refactors the TokenSelector to handle
pagination by using PaginatedCollections wrapped in PaginatedViews.

refs CNVS-1170

test-plan:
 - full regression on conversation recipient handling, inbox filter
   handling, section searching in the user's tab of course settings,
   etc. basically anywhere the recipient finder style input is used

 - in all those places, functionality should be equivalent with the
   existing code. may look slightly different, but handle the same and
   without looking bad

 - in addition, when searching for just users (I think this may only
   apply in the quick start bar finders), should handle multiple pages
   with automatic on-demand pagination (i.e. as you scroll down, more
   results load).

Change-Id: Ic0125b8f8452ecdb821802b0c415b4c6318e7c75
Reviewed-on: https://gerrit.instructure.com/18247
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>