optional merge proc for BookmarkedCollection.merge
when provided, duplicates across collections will be collapsed. only the
first instance of an item mapping to a given bookmark will be included,
but for each duplicate, the kept element and the duplicate will be
yielded to the provided block. the block should perform any necessary
merging of the elements so the kept element ends up with all information
from any of the merged elements.

refs CNVS-1170

test-plan:
 - create two BookmarkCollections using scopes and bookmarkers that will
   produce duplicates across the collections
 - merge them with BookmarkCollection.merge(...) and NO block
 - paginating the merged collection should include all elements from
   each collection (properly sorted) with no deduplication
 - merge them with BookmarkCollection.merge(...) and a block
 - paginating the merged collection should include elements from either
   collection (properly sorted) only once

Change-Id: I1cb40b657e0db88d9cb4fa0f7f2d1341cdd71eb5
Reviewed-on: https://gerrit.instructure.com/18248
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>