update Japanese translation to prevent error pages
Some of the countable translation keys were not (and still are not)
correct in the translation file. This update will prevent Canvas
from giving page errors when the translation keys are missing.
(meanwhile, looking in to the incorrect countables in translation)

Test Plan:
  - Canvas pages with countable translations (e.g. "1 student" vs
    "3 students") should not be a page error.

fixes CNVS-4533

Change-Id: I6c076805c590f13b1381b8d870a943f80077e835
Reviewed-on: https://gerrit.instructure.com/18563
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>