spec: fix teacher speed grader spec fix
wraps clicking the link to the gradebook in a "expect_new_page_load"
call in order to avoid race conditions for elements to become stale.

Also adds an ID to the gradebook link in speed grader because the
previous selector was based on an element with two links, this stops the
chance of Selenium selecting the wrong list inadvertently causing the
spec to fail.

test plan:
  - specs pass.

Change-Id: If6cd81ac3a801fcbab1a00814d6dac32bf670717
Reviewed-on: https://gerrit.instructure.com/18748
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>