record last_inline_view time for attachments
to collect data about when scribd (or crocodoc) documents
are not being used and can be purged from third party servers
to save money

fixes CNVS-5310

note: this data is not yet exposed; it can only be queried
from the database manually for now

test plan:  make sure the last_inline_view attribute for an
Attachment is updated whenever the attachment is viewed via
scribd or crocodoc. this can be done in several ways:
 - via the preview icon that appears next to a document link
   in rich text
 - on the files page
 - for student submissions on SpeedGrader

the easiest way to check this is in the Rails console.
get the attachment id from the canvas url, then,
in the console (after switching to the correct shard),
do

  Attachment.find(id).last_inline_view

Change-Id: I137ab631ad4eefdf4d9af353dd11191b7f2ca140
Reviewed-on: https://gerrit.instructure.com/20810
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: James Williams  <jamesw@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>
QA-Review: Bracken Mosbacker <bracken@instructure.com>