don't silently eat failure in cassandra page views
if recording a page view fails, we want to handle and stop the failure
so it doesn't fail out the whole batch. but we don't want to do so
silently; log the error instead. (in the case of cassandra page views,
this failure will actually be a failure to log the failure in the
event_stream_failures table.)

fixes CNVS-6199

test-plan:
  - turn off cassandra for page views but leave redis queueing on
  - queue a page view (via script/console) and note its request id
  - queue a second page view with the same request id
  - queue a third page view with a different request id
  - process the queue (PageView.process_cache_queue)
  - the first and third page views should be inserted into the database
  - there should be crit-level lines in the log about the failure of the
    second page view

Change-Id: Ie1849a73d8f625ab8f90c8c15ddff5f9150a0de4
Reviewed-on: https://gerrit.instructure.com/21263
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>