add an on_load callback to simply_versioned
refs CNVS-6869

When a version is loaded from the database, the on_load callback
gives the model an opportunity to process or modify the data if
necessary. For example, if current information on the model should
override a cached attribute, this would be the place to do it.

Test Plan:
- in the console, create a submission, e.g.
  >> s = Submission.create :assignment => a1, :user => u1
- add an on_load callback, e.g.
  >> submission.simply_versioned_options[:on_load] =
     lambda{ |model, version| model.grade = 'A' }
- now, no matter what grade you set (and save) in a versioned copy of
  the submission, you should get 'A' back

Change-Id: I9dde0118e2833c72e6209f3d632d9503db1e1eb3
Reviewed-on: https://gerrit.instructure.com/22472
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Product-Review: Duane Johnson <duane@instructure.com>