fix PaginatedView performance issue/bug, fixes CNVS-7309
fix an issue introduced in c95537ba where the same fetch can be kicked off
many times. backbone removes dups, so it's not a huge deal, but it is
suboptimal behavior. for PageViewCollection, however, this results in
duplicate models in the collection (since the model has no id).

because we do a fetchNextPageIfNeeded on any collection event, and
multiple events can be triggered in the same tick (e.g. reset + add),
several of the same fetch will happen (fetchingNextPage is false when we
check it for all of them, but it's true after the first setTimeout runs).

test plan:
1. open your chrome console
2. under settings, check the "Log XMLHttpRequests" box
3. enable page views and go to a user page
4. if you have fewer than 100 page views, lower per_page so that paginated
   requests happen
5. you should only see one of each ajax request in the logs
6. you should not see duplicate/out of order records in the UI

Change-Id: I11fe68c6955bf42fc3b4452b45bb441b1ff0f87b
Reviewed-on: https://gerrit.instructure.com/22964
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Product-Review: Marc LeGendre <marc@instructure.com>
QA-Review: Marc LeGendre <marc@instructure.com>