request blacklist based on user id, access token, or ip
refs CNVS-7008

This middleware reads and caches a blacklist from a Setting, then
compares it against the request (access token || user id) and the
request IP address.

You can change the blacklist and SIGHUP the rails process to reload it
without restarting rails.

Note that because of the limitations of middleware in rails 2.3 apps,
this doesn't look at the access token if it's given in a POST params
body, rather than the HTTP header or query string params.

test plan:

* By default this middleware won't block anybody. Verify that you don't
  get 403 errors making canvas requests.
* The blacklist is a comma-separated list of identifiers. To set it:
  Setting.set('request_throttle.blacklist', '1234') and then reload
  Canvas.
* Add your user global_id to the blacklist, verify you are blocked in
  web ui requests.
* Add an access token to the blacklist and use it, verify you are
  blocked.
* Add your IP to the blacklist (likely "127.0.0.1" for local dev envs),
  and verify you are blocked.
* Verify others are not blocked after these adds.

Change-Id: I4a1c7ef4bab70c586d603dc01f3ca440d211f11d
Reviewed-on: https://gerrit.instructure.com/23293
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>