add prereqs pill to draft state modules page
Prereq pill now added to each module. If there are 0 prereqs, the pill
does not show up, if there is 1 prereq, the name of the prereq module
is displayed and if there are more than 1 prereq then the number of the
prereqs are displayed with their names being displayed in a tooltip
when hovering over with a mouse

Test Plan
1. Go to the modules page. Ensure there are modules with 0, 1 and 2
prereqs created.
2. Ensure draft state is enabled.
3. The module with 0 prereqs should not display any prereqs "pill". The
module with 1 prereq should display the name of the prereq module in a
pill and the module with 2 prereqs should show the number "2" with a
tool tip displaying the names of the 2 modules when you hover over the
"pill"

fixes CNVS-7995

Change-Id: I07d0d2282a0fcbc6265e2d082e0747ef1a8938a5
Reviewed-on: https://gerrit.instructure.com/24141
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>